Add files via upload
We now ran a macro to determine the starting price of DQ and the ending price in 2018 in order to measure how much the stock raised or dropped. We borrowed code from https://stackoverflow.com/questions/18088729/row-count-where-data-exists in order to accurately define how to determine the first row of data for our sample (our starting price), then doctored the code to find our ending price.